Refactor to only support modality layout
