+++ src/DIPS.Xamarin.UI/Internal/Xaml/FloatingActionMenu.xaml
@@ -8,11 +8,10 @@
              xmlns:ValueConverters="clr-namespace:DIPS.Xamarin.UI.Converters.ValueConverters"
              mc:Ignorable="d"
              x:Name="CodeBehind"
-             x:Class="DIPS.Xamarin.UI.Controls.FloatingActionMenu.FloatingActionMenu">
+             x:Class="DIPS.Xamarin.UI.Internal.Xaml.FloatingActionMenu">
 
     <ContentView.Content>                                   
         <Button x:Name="ExpandButton"
-                BindingContext="{x:Reference CodeBehind}"
                 Opacity=".4"
                 WidthRequest="{Binding Size}"
                 HeightRequest="{Binding Size}"
