+++ src/Samples/DIPS.Xamarin.UI.Samples/Controls/FloatingActionMenu/FloatingActionMenuPage.xaml
@@ -12,29 +12,23 @@
     <ContentPage.Content>
         <dxui:ModalityLayout>
             <dxui:ModalityLayout.Behaviors>
-                <dxui:FloatingActionMenuBehaviour>
-                    <dxui:FloatingActionMenu Size="60"
-                                             ExpandButtonBackgroundColor="CornflowerBlue"
-                                             ExpandButtonText="Icon"
-                                             XPosition=".75"
-                                             YConstraint=".85">
-                        <dxui:MenuButton BackgroundColor="LightSeaGreen"
-                                         Text="A"
-                                         TextColor="White"
-                                         Title="First page" />
-                        <dxui:MenuButton BackgroundColor="LightSeaGreen"
-                                         Text="A"
-                                         TextColor="White"
-                                         Title="Second page" />
-                        <dxui:MenuButton BackgroundColor="LightSeaGreen"
-                                         Text="A"
-                                         TextColor="White"
-                                         Title="Third page" />
-                        <dxui:MenuButton BackgroundColor="LightSeaGreen"
-                                         Text="A"
-                                         TextColor="White"
-                                         Title="Fourth page" />
-                    </dxui:FloatingActionMenu>
+                <dxui:FloatingActionMenuBehaviour Size="60">
+                    <dxui:MenuButton BackgroundColor="Aqua"
+                                     Text="A"
+                                     TextColor="White"
+                                     Title="First page" />
+                    <dxui:MenuButton BackgroundColor="LightSeaGreen"
+                                     Text="A"
+                                     TextColor="White"
+                                     Title="Second page" />
+                    <dxui:MenuButton BackgroundColor="LightSeaGreen"
+                                     Text="A"
+                                     TextColor="White"
+                                     Title="Third page" />
+                    <dxui:MenuButton BackgroundColor="LightSeaGreen"
+                                     Text="A"
+                                     TextColor="White"
+                                     Title="Fourth page" />
                 </dxui:FloatingActionMenuBehaviour>
             </dxui:ModalityLayout.Behaviors>
         </dxui:ModalityLayout>
